requirements.txt
anthropic>=0.40.0
google-genai>=1.0.0
httpx>=0.27.0
